server/requirements.txt
Flask==3.0.3
google-auth==2.16.0
google-api-python-client==2.53.0
google-auth==2.35.0
google-cloud-resource-manager
python-dotenv
flask-cors
